+++ src/cssLanguageTypes.ts
@@ -5,6 +5,7 @@
 'use strict';
 
 import { Range, TextEdit, Position } from "vscode-languageserver-types";
+import { IEntryData } from "./services/languageFacts";
 
 export { Range, TextEdit, Position };
 
@@ -47,4 +48,11 @@ export interface ICompletionParticipant {
 
 export interface DocumentContext {
 	resolveReference(ref: string, base?: string): string;
+}
+
+export interface LanguageServiceOptions {
+	customProperties?: IEntryData[];
+	customAtDirectives?: IEntryData[];
+	customPseudoClasses?: IEntryData[];
+	customPseudoElements?: IEntryData[];
 }
